updated int to long

+++ threatconnect-app/apps-core/src/main/java/com/threatconnect/app/apps/AppConfig.java
@@ -121,9 +121,9 @@ public abstract class AppConfig
 		return getString(TC_API_PATH);
 	}
 	
-	public Integer getTcUserId()
+	public Long getTcUserId()
 	{
-		return getInteger(TC_USER_ID);
+		return getLong(TC_USER_ID);
 	}
 	
 	public Integer getTcSpaceElementId()
